fix: "urlencodedParser is not iterable" error
The express loader pulls middleware out of the stack to reorder it, but
it wasn't correctly handling them when they were missing. If the
middleware type was not found the variable was left undefined and
resulted in an error when trying to use the spread syntax on it later.
Initialize each of these to an empty array so we have a no-op fallback
if any of the middleware are not found.